Made some changes to project structure. Added steam id linking
