+++ .gitignore
@@ -9,3 +9,4 @@
 /api/mail.bat
 /api/test.bat
 /api/test.py
+/.env
