+++ api/srlm/__init__.py
@@ -1,7 +1,6 @@
 import api.srlm.definitions
 import api.srlm.app
 import api.srlm.logger
-import api.srlm.exceptions
 import api.srlm.api_access
 
 from api.srlm.run import asgi_app as asgi
