+++ api/srlm/app/api/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint('api', __name__)
 
-from api.srlm.app.api import errors, players, tokens, users, auth, permissions, divisions, leagues, seasons, teams, season_divisions, match, test
+from api.srlm.app.api import auth, users, league, utils, games
