+++ api/srlm/app/api/auth/routes.py
@@ -2,31 +2,33 @@ from flask import request, url_for
 from api.srlm.api_access.models import AuthorizedApp
 from api.srlm.app import db
 from api.srlm.app.api import bp
-from api.srlm.app.api.auth import basic_auth, req_app_token, user_auth
-from api.srlm.app.api.responses import request_success
-from api.srlm.app.auth.functions import get_bearer_token
+from api.srlm.app.api.auth.utils import basic_auth, req_app_token, user_auth
+from api.srlm.app.api.utils import responses
+from api.srlm.app.api.auth.utils import get_bearer_token
 from api.srlm.app.models import User
 
 
-@bp.route('/tokens/user', methods=['POST'])
+@bp.route('/user', methods=['POST'])
 @basic_auth.login_required
 def get_user_token():
+    """Requests an auth token for a user."""
     token = basic_auth.current_user().get_token()
     expires = basic_auth.current_user().token_expiration
     db.session.commit()
     return {'token': token, 'expires': expires}
 
 
-@bp.route('/tokens/user', methods=['DELETE'])
+@bp.route('/user', methods=['DELETE'])
 @req_app_token
 @user_auth.login_required
 def revoke_user_token():
+    """Revokes the users current auth token"""
     user_auth.current_user().revoke_token()
     db.session.commit()
-    return request_success('User token revoked')
+    return responses.request_success('User token revoked')
 
 
-@bp.route('/tokens/user/validate', methods=['POST'])
+@bp.route('/user/validate', methods=['POST'])
 @req_app_token
 @user_auth.login_required
 def validate_user_token():
@@ -43,7 +45,7 @@ def validate_user_token():
     return response
 
 
-@bp.route('/tokens/app', methods=['POST'])
+@bp.route('/app', methods=['POST'])
 @req_app_token
 def request_new_app_token():
     app_token = get_bearer_token(request.headers)['app']
@@ -58,7 +60,7 @@ def request_new_app_token():
     return response
 
 
-@bp.route('/tokens/app', methods=['GET'])
+@bp.route('/app', methods=['GET'])
 @req_app_token
 def get_app_token():
     app_token = get_bearer_token(request.headers)['app']
