+++ api/srlm/app/api/games/__init__.py
@@ -1,0 +1,1 @@
+from api.srlm.app.api.games import lobby, match
