+++ api/srlm/app/api/games/lobby.py
@@ -1,8 +1,7 @@
 from flask import request
-
 from api.srlm.app.api import bp
-from api.srlm.app.api.auth import req_app_token
-from api.srlm.app.api.functions import force_fields, ensure_exists
+from api.srlm.app.api.auth.utils import req_app_token
+from api.srlm.app.api.utils.functions import force_fields, ensure_exists
 from api.srlm.app.models import Match
 from api.srlm.app.spapi.lobby_manager import generate_lobby
