+++ api/srlm/app/api/league/__init__.py
@@ -1,0 +1,1 @@
+from api.srlm.app.api.league import leagues, seasons, divisions, season_divisions, players, teams
