+++ api/srlm/app/api/league/players.py
@@ -1,15 +1,12 @@
 from datetime import datetime, timezone
-
 from flask import request
 import sqlalchemy as sa
-
 from api.srlm.app import db
-from api.srlm.app.api import bp, responses
-from api.srlm.app.api.auth import req_app_token
-
-# create a new logger for this module
-from api.srlm.app.api.errors import BadRequest, ResourceNotFound
-from api.srlm.app.api.functions import ensure_exists, force_fields, force_unique, clean_data
+from api.srlm.app.api import bp
+from api.srlm.app.api.utils import responses
+from api.srlm.app.api.auth.utils import req_app_token
+from api.srlm.app.api.utils.errors import BadRequest, ResourceNotFound
+from api.srlm.app.api.utils.functions import ensure_exists, force_fields, force_unique, clean_data
 from api.srlm.app.models import Player, SeasonDivision, Team, PlayerTeam, FreeAgent
 from api.srlm.logger import get_logger
 log = get_logger(__name__)
