+++ api/srlm/app/api/league/teams.py
@@ -1,14 +1,13 @@
 from datetime import datetime, timezone
-
 from api.srlm.app import db
-from api.srlm.app.api import bp, responses
+from api.srlm.app.api import bp
+from api.srlm.app.api.utils import responses
 from flask import request, url_for
 import sqlalchemy as sa
-
-from api.srlm.app.api.errors import ResourceNotFound, BadRequest
-from api.srlm.app.api.functions import ensure_exists, force_fields, force_unique, clean_data
+from api.srlm.app.api.utils.errors import ResourceNotFound, BadRequest
+from api.srlm.app.api.utils.functions import ensure_exists, force_fields, force_unique, clean_data
 from api.srlm.app.models import Team, SeasonDivision, PlayerTeam
-from api.srlm.app.api.auth import req_app_token
+from api.srlm.app.api.auth.utils import req_app_token
 
 # create a new logger for this module
 from api.srlm.logger import get_logger
