--- api/srlm/app/auth/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint('auth', __name__)
-
-from api.srlm.app.auth import email, functions
