+++ api/srlm/app/email.py
@@ -1,10 +1,14 @@
-from flask import current_app
+import os
+from flask import current_app, render_template
 from flask_mail import Message
 from threading import Thread
 from api.srlm.app import mail
 
 # TODO CONVERT TO CELERY
 # Sending mail with the context of the app
+from api.srlm.definitions import app_name
+
+
 def send_async(app, msg):
     with app.app_context():
         mail.send(msg)
@@ -18,3 +22,26 @@ def send(subject, sender, recipients, text_body, html_body):
     # Requires access to the protected member '_get_current_object()' as
     # current_app is a proxy and has no value outside of the context of the main thread
     Thread(target=send_async, args=(current_app._get_current_object(), msg)).start()
+
+
+def send_password_reset_email(user):  # TODO change
+    token = user.get_password_reset_token()
+    print(token)
+    send(f'{app_name} - Reset Your Password',
+         sender=os.getenv('OUTGOING_MAIL_ADDRESS'),
+         recipients=[user.email],
+         text_body=render_template('auth/reset_password_email.txt',
+                                   user=user, token=token),
+         html_body=render_template('auth/reset_password_email.html',
+                                   user=user, token=token))
+
+
+def send_new_user_email(user):  # TODO change
+    token = user.get_password_reset_token()
+    send(f'{app_name} - Account creation',
+         sender=os.getenv('OUTGOING_MAIL_ADDRESS'),
+         recipients=[user.email],
+         text_body=render_template('auth/new_user_email.txt',
+                                   user=user, token=token, domain=current_app.config['DOMAIN']),
+         html_body=render_template('auth/new_user_email.html',
+                                   user=user, token=token, domain=current_app.config['DOMAIN']))
