+++ requirements.txt
@@ -1,6 +1,7 @@
 aiosmtpd==1.4.5
 alembic==1.10.3
 amqp==5.2.0
+apispec==6.6.0
 asgiref==3.8.0
 async-timeout==4.0.3
 atpublic==4.0
@@ -18,12 +19,14 @@ colorama==0.4.6
 defusedxml==0.7.1
 dnspython==2.3.0
 email-validator==2.0.0.post1
+eventlet==0.36.1
 exceptiongroup==1.2.0
 Flask==2.2.3
 Flask-CeleryExt==0.5.0
 Flask-HTTPAuth==4.8.0
 Flask-Login==0.6.2
 Flask-Mail==0.9.1
+flask-marshmallow==1.2.1
 Flask-Migrate==4.0.4
 Flask-Script==2.0.6
 Flask-SQLAlchemy==3.0.3
@@ -42,11 +45,15 @@ importlib-metadata==6.4.1
 itsdangerous==2.1.2
 Jinja2==3.1.2
 kombu==5.3.6
+ma==1.0.2
+ma.mongo==1.0.2
 Mako==1.2.4
 markdown-it-py==3.0.0
 MarkupSafe==2.1.2
+marshmallow==3.21.1
 mdurl==0.1.2
 multidict==6.0.5
+packaging==24.0
 priority==2.0.0
 prometheus_client==0.20.0
 prompt-toolkit==3.0.43
@@ -71,6 +78,7 @@ tzdata==2024.1
 urllib3==2.2.1
 vine==5.1.0
 wcwidth==0.2.13
+webargs==8.4.0
 Werkzeug==2.2.3
 wsproto==1.2.0
 WTForms==3.0.1
